refactor: ファイル分割の変更
端末を判定する部分を components から外すため
