+++ src/App.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from "react";
 import PrefectureList from "./components/PrefecturesList";
 import Graph from "./components/graph/Graph";
 import Title from "./components/Title";
-import { useMediaQueryContext } from "./components/responsive/MediaQueryProvider";
+import { useMediaQueryContext } from "./responsive/MediaQueryProvider";
 
 const App = () => {
   const [graphData, setGraphData] = useState<object[]>();
